+++ lib/View/MedicalCalculatorScreen.dart
@@ -8,15 +8,15 @@ import 'package:url_launcher/url_launcher.dart';
 
 import 'LoginScreen.dart';
 
-class MedicalCalculatorView extends StatefulWidget {
-  const MedicalCalculatorView({Key? key}) : super(key: key);
+class MedicalCalculatorScreen extends StatefulWidget {
+  const MedicalCalculatorScreen({Key? key}) : super(key: key);
   static String id="medicalCalc";
 
   @override
-  State<MedicalCalculatorView> createState() => _MedicalCalculatorViewState();
+  State<MedicalCalculatorScreen> createState() => _MedicalCalculatorScreenState();
 }
 
-class _MedicalCalculatorViewState extends State<MedicalCalculatorView> {
+class _MedicalCalculatorScreenState extends State<MedicalCalculatorScreen> {
 
   bool search = false;
   bool searchCriteria=true;
@@ -77,7 +77,7 @@ class _MedicalCalculatorViewState extends State<MedicalCalculatorView> {
                         hintStyle: TextStyle(color: Colors.grey),
                         prefixIcon: Icon(
                           Icons.search,
-                          color: Colors.blue.shade900,
+                          color: Colors.indigo.shade900,
                         ),
                         fillColor: Colors.white,
                         contentPadding: const EdgeInsets.symmetric(
@@ -94,7 +94,7 @@ class _MedicalCalculatorViewState extends State<MedicalCalculatorView> {
             : Center(
           child: Text(
             'Medical Calculators',
-            style: TextStyle(color: Colors.blue.shade900),
+            style: TextStyle(color: Colors.indigo.shade900),
             textAlign: TextAlign.left,
           ),
         ),
@@ -109,7 +109,7 @@ class _MedicalCalculatorViewState extends State<MedicalCalculatorView> {
                 margin: EdgeInsets.only(right: 16),
                 child: Icon(
                   search?Icons.search_off:Icons.search,
-                  color: Colors.blue.shade900,
+                  color: Colors.indigo.shade900,
                 )),
             onTap: () {
               setState(() {
@@ -123,16 +123,19 @@ class _MedicalCalculatorViewState extends State<MedicalCalculatorView> {
                 margin: EdgeInsets.only(right: 16),
                 child: Icon(
                   Icons.output_sharp,
-                  color: Colors.blue.shade900,
+                  color: Colors.indigo.shade900,
                 )),
             onTap: () async{
               final auth = FirebaseAuth.instance;
 
-              auth.signOut();
-SharedPreferences preferences = await SharedPreferences.getInstance();
-await preferences.clear();
-              Navigator.pushNamedAndRemoveUntil(
-                  context, LoginScreen.id, (route) => false);
+              final ConfirmAction action = (await _asyncConfirmDialog(context))!;
+
+              if(action==ConfirmAction.Accept){
+                auth.signOut();
+                SharedPreferences preferences = await SharedPreferences.getInstance();
+                await preferences.clear();
+                Navigator.pushNamedAndRemoveUntil(context, LoginScreen.id, (route) => false);}
+
             },
           ),
         ],
@@ -145,7 +148,7 @@ await preferences.clear();
           slivers: [
             SliverToBoxAdapter(
               child: Container(
-                margin: const EdgeInsets.only(top: 8),
+                margin: const EdgeInsets.all(8),
                 color: Colors.grey.shade200,
                 // padding: EdgeInsets.all(16),
                 height: MediaQuery.of(context).size.height*0.9,
@@ -475,3 +478,35 @@ await preferences.clear();
 
   }
 }
+
+
+enum ConfirmAction { Cancel, Accept}
+Future<ConfirmAction?> _asyncConfirmDialog(BuildContext context) async {
+  return showDialog<ConfirmAction>(
+    context: context,
+    barrierDismissible: false, // user must tap button for close dialog!
+    builder: (BuildContext context) {
+      return AlertDialog(
+        title: Text('Sign out'),
+        content: const Text(
+            'Are you sure you want to sign out?'),
+        actions: <Widget>[
+          ElevatedButton(
+
+            child: const Text('Cancel'),
+            onPressed: () {
+              Navigator.of(context).pop(ConfirmAction.Cancel);
+            },
+          ),
+          ElevatedButton(
+            child: const Text('Accept'),
+            onPressed: () {
+              Navigator.of(context).pop(ConfirmAction.Accept);
+            },
+          )
+        ],
+      );
+    },
+  );
+}
+
